more stuff out of init
