+++ joplin/base/models/day_and_duration.py
@@ -1,0 +1,36 @@
+from django.db import models
+
+from modelcluster.models import ClusterableModel
+
+from wagtail.admin.edit_handlers import FieldPanel
+
+class DayAndDuration(ClusterableModel):
+    MONDAY = 'Monday'
+    TUESDAY = 'Tuesday'
+    WEDNESDAY = 'Wednesday'
+    THURSDAY = 'Thursday'
+    FRIDAY = 'Friday'
+    SATURDAY = 'Saturday'
+    SUNDAY = 'Sunday'
+    DAY_OF_WEEK_CHOICES = (
+        (MONDAY, 'Monday'),
+        (TUESDAY, 'Tuesday'),
+        (WEDNESDAY, 'Wednesday'),
+        (THURSDAY, 'Thursday'),
+        (FRIDAY, 'Friday'),
+        (SATURDAY, 'Saturday'),
+        (SUNDAY, 'Sunday'),
+    )
+
+    day_of_week = models.CharField(max_length=20, choices=DAY_OF_WEEK_CHOICES)
+    start_time = models.TimeField()
+    end_time = models.TimeField()
+
+    content_panels = [
+        FieldPanel('day_of_week'),
+        FieldPanel('start_time'),
+        FieldPanel('end_time'),
+    ]
+
+    def __str__(self):
+        return f'{self.day_of_week} {self.start_time} - {self.end_time}'
